Moved glass button to its own class.
